1.19

+++ src/main/java/net/boster/particles/main/particle/dust/DustOptionsProvider.java
@@ -4,5 +4,5 @@ import org.jetbrains.annotations.NotNull;
 
 public interface DustOptionsProvider {
 
-    @NotNull BPDustOptions create(int r, int g, int b, int amount);
+    @NotNull BPDustOptions create(int r, int g, int b, float amount);
 }
