+++ src/main/java/net/boster/particles/main/particle/dust/OldDustOptionsProvider.java
@@ -5,7 +5,7 @@ import org.jetbrains.annotations.NotNull;
 public class OldDustOptionsProvider implements DustOptionsProvider {
 
     @Override
-    public @NotNull BPDustOptions create(int r, int g, int b, int amount) {
+    public @NotNull BPDustOptions create(int r, int g, int b, float amount) {
         return () -> new int[]{r, g, b};
     }
 }
